don't special-case nullary constructors
This is upsetting, but there's no way for Foo.Bar to work *and* for Foo.Bar()
to work without sacrificing instanceof. That is, there's no way to make
`Foo.Bar() instanceof Foo.Bar` work, as far as I can figure out.

Another option that I may pursue is to special-case nullary constructors even
further, and make them not be constructors at all, just values. That...
might be the right way to go, but it would not play as nicely with matchbook.
Though matchbook should allow it...